% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq_download.R
\name{aq_download}
\alias{aq_download}
\title{Get formatted timeseries data from Aquarius.}
\usage{
aq_download(
  loc_id,
  ts_name,
  start = "1950-01-01",
  end = as.character(Sys.Date()),
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{loc_id}{The location ID, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \verb{29EA001} or \code{YOWN-0804}.}

\item{ts_name}{The timeseries name, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \code{Wlevel_bgs.Calculated}.}

\item{start}{The first day for which you want information (local time). Whole days only. Times requested prior to the actual timeseries start will be adjusted to match available data.}

\item{end}{The last day for which you want information (local time). Whole days only. Times requested prior to the actual timeseries end will be adjusted to match available data.}

\item{login}{Your Aquarius login credentials as a character vector of two. Default pulls information from your .renviron profile; see details.}

\item{server}{The URL for your organization's Aquarius web server. Default for the Yukon Water Resources Branch.}
}
\value{
A list with four data.frames: station metadata; timeseries information consisting of timestamps, values, applicable grade and approval levels; approval level change summary; grade level change summary. Important: all times in this list are UTC.
}
\description{
Fetches and processes data downloaded from an Aquarius web-hosted server and returns it in a concise format. Note that any times returned by this function are in UTC.
}
\details{
To store login credentials in your .renviron profile, call usethis::edit_r_environ() and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
}
