% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq_upload.R
\name{aq_upload}
\alias{aq_upload}
\title{Upload data to Aquarius}
\usage{
aq_upload(
  loc_id,
  ts_name,
  data,
  start = NULL,
  end = NULL,
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{loc_id}{The location ID, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \verb{29EA001} or \code{YOWN-0804}.}

\item{ts_name}{The timeseries name, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \code{Wlevel_bgs.Calculated}.}

\item{data}{The data you wish to append to an existing timeseries. Must contain columns named \code{Value} and \code{Time}. See details for more information.}

\item{start}{Not required; specify only if you with to overwrite existing data, as a POSIXct object. Inclusive, must be >= to the first data point in \code{data}. Timezones should match that of data$Time.}

\item{end}{Not required; specify only if you with to overwrite existing data, as a POSIXct object. Inclusive, must be <= to the final data point in \code{data}. Timezones should match that of data$Time.}

\item{login}{Your Aquarius login credentials as a character vector of two. Default pulls information from your .renviron profile; see details.}

\item{server}{The URL for your organization's Aquarius web server. Default is for the Yukon Water Resources Branch.}
}
\value{
A list with four data.frames: station metadata; timeseries information consisting of timestamps, values, applicable grade and approval levels; approval level change summary; grade level change summary. Important: all times in this list are in UTC.
}
\description{
This function bypasses the web GUI and allows you to append data to Aquarius directly.
}
\details{
The parameter \code{data} should consist of a data.frame with two named columns: Value and Time. Units for the column \code{Value} are set according the units already in use for the timeserie, so it should only contain numbers compatible with this. The \code{Time} column should be formatted as.POSIXct, with the timezone attribute correctly set. Failure to ensure the correct timezone of input data will result in offset points.

To store login credentials in your .renviron profile, call usethis::edit_r_environ() and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
}
