% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_get_timeseries.R
\name{DB_get_timeseries}
\alias{DB_get_timeseries}
\title{Get timeseries information from the hydromet database}
\usage{
DB_get_timeseries(
  path = "default",
  operator = NULL,
  location = NULL,
  type = NULL,
  parameter = NULL
)
}
\arguments{
\item{path}{The path to the database, passed to hydroConnect. Default uses hydroConnect default path.}

\item{operator}{Narrow by location operator if you wish. Exact spelling only!}

\item{location}{Narrow by location if you wish. Exact spelling only!}

\item{type}{Narrow by type if you wish. Exact spelling only!}

\item{parameter}{Narrow by parameter if you wish. Exact spelling only!}
}
\value{
A data.frame containing the database timeseries matching the function parameters.
}
\description{
Wondering what's in the database? This function helps you see what's under the hood, with an eye to helping you create a query for function get_DB_data. Leaving all NULL defaults will show you every timeseries in the database.
}
