% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_get_ts.R
\name{DB_get_ts}
\alias{DB_get_ts}
\title{Get timeseries data from the hydromet database}
\usage{
DB_get_ts(
  path = "default",
  location,
  parameter,
  frequency,
  start,
  end,
  save_path = "choose"
)
}
\arguments{
\item{path}{The path to the database, passed to \code{\link[=hydroConnect]{hydroConnect()}}. Default uses hydroConnect default path.}

\item{location}{The location code/id(s) for the requested timeseries as a character vector of 1 or more. Refer to DB_get_timeseries if unsure of location code.}

\item{parameter}{The parameter requested for the timeseries.  Refer to \code{\link[=DB_get_timeseries]{DB_get_timeseries()}} if unsure of parameter spelling.}

\item{frequency}{One of "daily", "realtime", or "discrete".}

\item{start}{The start date or datetime of records requested (inclusive). Specify a Date or poSIXCT object, or a character vector of form "2022-01-01" or "2022-01-01 10:10:10". Set before timeseries start to get all records up to the end date/time.}

\item{end}{The end date or datetime of records requested (inclusive). Format as per 'start'. Set after timeseries end to get all records after the start date/time}

\item{save_path}{Specify a path here if you want an Excel workbook saved to disk. "choose" lets you interactively choose your folder.}
}
\value{
A list of data frames containing the information requested and, optionally, an Excel workbook saved to disk.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Extracts timeseries information from the WRB's hydrometric database. Refer to \code{\link[=DB_browse_ts]{DB_browse_ts()}} if unsure of how to formulate your query.
}
\seealso{
\code{\link[=DB_browse_ts]{DB_browse_ts()}} for browsing timeseries data from the database, or, for spatial data, \code{\link[=DB_browse_spatial]{DB_browse_spatial()}} and \code{\link[=DB_get_spatial]{DB_get_spatial()}} to browse and extract spatial data. For location metadata use \code{\link[=DB_get_meta]{DB_get_meta()}}.
}
