% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroConnect.R
\name{hydroConnect}
\alias{hydroConnect}
\title{Connect to hydrology database}
\usage{
hydroConnect(path, timeout = 1e+05)
}
\arguments{
\item{path}{The path to the database. Currently supports only .sqlite databases.}

\item{timeout}{For sqlite DBs, the duration in which to retry an operation in milliseconds. Valid for the duration of the connection.}
}
\value{
A connection to the database.
}
\description{
This function exists to facilitate connecting to the hydrology database, especially if the database type and connection method changes in the future.
}
