% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometPlot.R
\name{hydrometPlot}
\alias{hydrometPlot}
\title{Multi-purpose plotting function}
\usage{
hydrometPlot(
  location,
  parameter,
  type = "continuous",
  startDay = 1,
  endDay = 365,
  tzone = "MST",
  years = NULL,
  datum = TRUE,
  title = TRUE,
  returns = "auto",
  return_type = "max",
  return_months = c(5:9),
  allowed_missing = 10,
  save_path = NULL,
  dbPath = "default"
)
}
\arguments{
\item{location}{The location for which you want a plot.}

\item{parameter}{The parameter you wish to plot. The location:parameter combo must be in the local database.}

\item{type}{The type of data you wish to plot. Select from "discrete" or "continuous". Again, must match what's in the database.}

\item{startDay}{The start day of year for the plot x-axis. Can be specified as a number from 1 to 365, as a character string of form "yyyy-mm-dd", or as a date object. Either way the day of year is the only portion used, specify years to plot under parameter \code{years}.}

\item{endDay}{The end day of year for the plot x-axis. As per \code{startDay}.}

\item{tzone}{The timezone to use for graphing. Only really evident for a small number of days.}

\item{years}{The years to plot. If \code{startDay} and \code{endDay} cover December 31 - January 1, select the December year(s). Max 10 years, NULL = current year.}

\item{datum}{Should a vertical datum be applied to the data, if available? TRUE or FALSE.}

\item{title}{Should a title be included?}

\item{returns}{Should returns be plotted? You have the option of using pre-determined level returns only (option "table"), auto-calculated values(option "calculate"), "auto" (priority to "table", fallback to "calculate"), or "none". Defaults to "auto". Only for \code{type = "continuous"}}

\item{return_type}{Use minimum ("min") or maximum ("max") values for returns?}

\item{return_months}{Numeric vector of months during which to look for minimum or maximum values. Only works with calculated returns. Does not have to be within \code{startDay} and \code{endDay}, but will only consider data up to the last year specified in \code{years}. For months overlapping the new year like November-April, should look like c(11:12,1:4). IMPORTANT: the first month in the range should be the first element of the vector: c(1:4, 11:12) would not be acceptable. Think of it as defining a season. Passed to 'months' argument of \code{\link[fasstr:calc_annual_extremes]{fasstr::calc_annual_extremes()}} and also used to set the 'water_year_start' parameter.}

\item{allowed_missing}{Allowable \% of data missing during the months specified in 'return_months' to still retain the year for analysis when calculating returns. Passed to 'allowed_missing' argument of \code{\link[fasstr:calc_annual_extremes]{fasstr::calc_annual_extremes()}}.}

\item{save_path}{Default is NULL and the graph will be visible in RStudio and can be assigned to an object. Option "choose" brings up the File Explorer for you to choose where to save the file, or you can also specify a save path directly.}

\item{dbPath}{The path to the local hydromet database, passed to \code{\link[=hydroConnect]{hydroConnect()}} or \code{\link[=DB_get_ts]{DB_get_ts()}}.}
}
\value{
A .png file of the plot requested (if a save path has been selected), plus the plot displayed in RStudio. Assign the function to a variable to also get a plot in your global environment as a ggplot object which can be further modified
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Intended to supersede \code{\link[WRBfloods:levelPlot]{WRBfloods::levelPlot()}} and \code{\link[WRBfloods:flowPlot]{WRBfloods::flowPlot()}}: see details.

Notice: in many cases, you're better off using the Shiny app at \code{\link[WRBfloods:hydroApp]{WRBfloods::hydroApp()}} to generate and export your plot. Read on if you need additional control over the final product.

This function plots data from the local hydrometric database (maintained by the WRBdatabase package) and yields consistent-looking plots for discrete and continuous data. This function can only plot what's in the database, so if you're unsure use the function \code{\link[=DB_browse_ts]{DB_browse_ts()}} to see what's in there first.

Plots of continuous-type data will include bands representing the historic min, max, and 25th-75th percentiles based on the last year of data plotted.

If necessary for performance, data may be down-sampled to daily means: see details.
}
\details{
Superseded function \code{\link[WRBfloods:levelPlot]{WRBfloods::levelPlot()}} and \code{\link[WRBfloods:flowPlot]{WRBfloods::flowPlot()}} can be used in the event that this function does not yield your desired graph, or if you cannot use the local database and must use data directly from the Water Survey of Canada.
\subsection{Down sampling of data}{

When requesting several years of data where high-frequency (realtime) data is available, some of the data may be displayed as daily means. In this case, the 100000 to 150000 points nearest to the spefied \code{endDate} will be loaded at maximum resolution while subsequent points will use daily means.
}

\subsection{Return period details:}{

Return periods generated using the "calculate" option are calculated using all available data up to and including the most recent plotted data. Extreme values are isolated for each year and passed to \code{\link[fasstr:compute_frequency_analysis]{fasstr::compute_frequency_analysis()}} using a Log Pearson Type 3 distribution and the method of moments to fit a curve to the distribution. Years (or rather periods within years; see below) missing more than the percentage of data specified in \code{allowed_missing} are excluded from consideration.

As the time of year during which minimum or maximum values happen and to prevent filtering out years that are missing irrelevant data (most rivers peak in spring, some peak in late summer, snow pack peaks in winter), the parameters return_months and allowed_missing should be carefully selected. This may require you to make a graph first to evaluate when the annual min/max of interest happens. Consider as well here that you can ask for return periods specific for a month of the year, such as return periods of minimum flows in September, or maximum flows in January. This also applies to levels or any other parameter which this function can plot.

If specifying the option "table", the function will attempt to find returns for the location within the package internal data. The returns contained there are from consultant work involving a detailed analysis of which data to keep/discard, as well as a human determination of the most appropriate distribution and curve fitting methods. The caveat here is that these tables are not necessarily up to date and that most locations in Yukon have no calculated return periods.
}
}
