% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowConnect.R
\name{snowConnect}
\alias{snowConnect}
\title{Connect to snow survey database}
\usage{
snowConnect(path)
}
\arguments{
\item{path}{The path to the database. Currently supports either a Microsoft Access database or an .sqlite database.}
}
\value{
A connection to the database.
}
\description{
If the Snow Survey database is migrated to an SQLite database, beware that the tables and table columns might no longer be what you're expecting!
}
