% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterInfo.R
\name{waterInfo}
\alias{waterInfo}
\title{Plots and tabular data for hydrometric locations}
\usage{
waterInfo(
  db_path = "//env-fs/env-data/corp/water/Common_GW_SW/Data/database/hydro.sqlite",
  locations = "all",
  level_flow = "both",
  end_date = Sys.Date(),
  months_min = c(1:4),
  months_max = c(4:9),
  allowed_missing = 10,
  save_path = "choose",
  plots = TRUE,
  plot_type = "combined",
  quiet = FALSE
)
}
\arguments{
\item{db_path}{The path to the local hydro database including extension. See WRBtools::hydroConnect for supported database types.}

\item{locations}{The list of locations requested, as either a vector of location IDs or one of "WRB" (only WRB stations selected), "WSC" (only), or "all". Default "all" fetches all stations.}

\item{level_flow}{Default 'both' will get and calculate level and flow information wherever possible. 'one' will pick flow where it exists, otherwise level. Exception to this is if there is no flow on the end_date requested AND most recent flow is > 1 month older than level ; in this case flow is assumed to be discontinued and level is used.}

\item{end_date}{The most recent day to include in calculations. Defaults to today.}

\item{months_min}{The month range in which to look for minimums.}

\item{months_max}{The month range in which to look for maximums.}

\item{allowed_missing}{The percent of data allowed to be missing and still return a minimum and/or maximum value. Applied separately to minimums and maximums for the months specified in 'months'.}

\item{save_path}{The path where the .csv(s) and plots should be saved.}

\item{plots}{Set TRUE if you want plots generated for each location (level or flow).}

\item{plot_type}{"combined" or "separate". Combined returns a single plot with yearly minimum and maximum on the same log-scale y axis, separate returns two stacked plots.}

\item{quiet}{Suppresses most messages and warnings.}
}
\value{
A list with four data.frames: location metadata, basic statistics, trend information, and daily measurements is returned to the R environment. In addition, an Excel workbook is saved to the save_path with the four data.frames as tabs, and a new folder created to hold level/flow plots for each station requested.
}
\description{
This function is intended to facilitate the reporting of hydrology data by compiling basic statistics (years of record, months of operation, min, max, etc.), trend information (Mann-Kendall direction and p-value, Sen's slope), and creating simple plots of level (for lakes) or flow for all requested locations. At its most basic (parameters to FALSE or NULL where applicable), the result is a list of two data.frames to the R environment with location metadata and field measurements.
}
