% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xle_convert.R
\name{xle_convert}
\alias{xle_convert}
\title{Convert Solinst logger files to csv format}
\usage{
xle_convert(
  xle_file = "choose",
  location,
  save_path = "choose",
  YOWN_master =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/1_YOWN_MASTER_TABLE/MASTER for R - remember to update.xlsx"
)
}
\arguments{
\item{xle_file}{The file you wish to convert. Default "choose" allows you to point to the file.}

\item{location}{The ID of the well in the form "YOWN-1500".}

\item{save_path}{The location where the csv file should be saved.}

\item{YOWN_master}{The location to the YOWN master spreadsheet.}
}
\value{
A csv of the logger data, ready for export to Aquarius or for general use.
}
\description{
Reads a Solinst .xle file and converts it into a .csv with proper column names. Converts units to those in common usage at the Yukon Water Resources Branch, standardizes file naming, and ensures that times are represented in UTC-7.
}
