% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWN_solinst_telem.R
\name{YOWN_solinst_telem}
\alias{YOWN_solinst_telem}
\title{YOWN Solinst LevelSender Telemetry Processing}
\usage{
YOWN_solinst_telem(active_telem = c("YOWN-2201S"))
}
\arguments{
\item{active_telem}{Character vector of YOWN sites for processing of SOLINST telemetry data}
}
\value{
Creates an .xle file in the XLE_File_Dropbox folder on the G drive
}
\description{
This script searches for new messages from the YOWN telemetry gmail inbox by specified sites with active telemetry, then downloads messages and combines each day's report into an XLE file which is then places in the XLE-dropbox folder for processing as usual. Trashed email upon successful creation of a .xle file
}
