% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRBtools-package.R
\docType{package}
\name{WRBtools-package}
\alias{WRBtools}
\alias{WRBtools-package}
\title{WRBtools: Utility functions for the Yukon Water Resources Branch}
\description{
Utility functions for use across the Yukon Water Resources Branch disciplines.
}
\author{
\strong{Maintainer}: Ghislain de Laplante \email{ghislaindel@gmail.com} (\href{https://orcid.org/0000-0002-5093-9185}{ORCID})

}
\keyword{internal}
